Fix #14: Room 109 and 20C Tyrant encounters cause camera glitches and soft lock
Disable Tyrant encounters for rooms 109, 20C and 117 as they cause the camera angle to glitch in the room. For 109 in particular, it gliches the camera angle in the connected rooms and blocks one of the doors which may cause a soft lock.
Room 117 disables a door and creates a new one, this can mess with the door rando, so just disable the encounter.