Fix #261: RE 2, random BGM requires RE1 installation (#262)
Before generating a seed, there is a check to see if the other game's BGM is selected, and if it is and that game is not installed, then the seed will not be generated and that error message will be displayed.

However, this check doesn't actually check which game BGMs are actually selected, it just checks if that game's name (either RE1 or RE2) is present in a hash set composed of RE1, RE2 and the names of any BGM folders. RE1 and RE2 are always present in this hash set so it will always think that you selected RE1 or RE2.

Fix it by checking RandoConfig.EnabledBGMs, to see if the BGM is actually selected.